### 2018-09-18 b6098e1
Update DOI in references
