+++ paper/paper.bib
@@ -37,7 +37,7 @@
   journal={Nature Methods},
   year={2018},
   publisher={Nature Publishing Group},
-  doi={doi.org/10.1038/s41592-018-0109-9}
+  doi={10.1038/s41592-018-0109-9}
 }
 
 @article{Katz2018,
