### 2018-09-18 be42a27
Updated formatting for JOSS .pdf compiling
