+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @article{ManteSussillo2013,
   title={Context-dependent computation by recurrent dynamics in prefrontal cortex},
   author={Mante, Valerio and Sussillo, David and Shenoy, Krishna V and Newsome, William T},
-  journal={nature},
+  journal={Nature},
   volume={503},
   number={7474},
   pages={78},
@@ -23,7 +23,7 @@
 @article{SussilloBarak2013,
   title={Opening the black box: low-dimensional dynamics in high-dimensional recurrent neural networks},
   author={Sussillo, David and Barak, Omri},
-  journal={Neural computation},
+  journal={Neural Computation},
   volume={25},
   number={3},
   pages={626--649},
@@ -37,13 +37,13 @@
   journal={Nature Methods},
   year={2018},
   publisher={Nature Publishing Group},
-  doi={10.1038/s41592-018-0109-9}
+  DOI={10.1038/s41592-018-0109-9}
 }
 
 @article{Katz2018,
   title={Using directional fibers to locate fixed points of recurrent neural networks},
   author={Katz, Garrett E and Reggia, James A},
-  journal={IEEE transactions on neural networks and learning systems},
+  journal={IEEE Transactions on Neural Networks and Learning Systems},
   volume={29},
   number={8},
   pages={3636--3646},
