### 2018-09-18 26e4b11
Testing both figures and captions; "availability"
