+++ paper/paper.md
@@ -32,17 +32,25 @@ Recurrent neural networks (RNNs) are powerful function approximators that can be
 
 Perhaps the most critical step in this reverse-engineering is identifying the fixed points of the RNN's internal dynamics. These fixed points, along with local linearizations of the RNN dynamics about those fixed points, provide a mechanistic description of how the network implements a computation. Identifying the fixed points of an RNN requires optimization tools that have previously been tailored for specific, "vanilla" or Hopfield-like RNN architectures [@SussilloBarak2013, @Katz2018]. These approaches rely on hard-coded analytic derivatives (e.g., of the hyperbolic tangent function) and thus can be cumbersome for complicated models (e.g., gated architectures), for studies that involve multiple different models, or when frequent architectural changes are required during the development of a model. While numerical derivatives can partially alleviate these difficulties, the additional computational costs typically make them infeasible for anything but the smallest of models.
 
-Here, we introduce [``FixedPointFinder``](https://github.com/mattgolub/fixed-point-finder), an open-source Tensorflow toolbox for finding fixed points and linearized dynamics in arbitrary RNN architectures. ``FixedPointFinder`` leverages Tensorflow's efficient automatic differentiation back-end to perform the optimizations and Jacobian calculations required for fixed-point identification and characterization. ``FixedPointFinder`` works with any arbitrary RNN architecture that conforms to Tensorflow's ``RNNCell`` API, including popular gated architectures (e.g., long-short term memory (LSTM) or gated recurrent units (GRU)), traditional "vanilla" architectures, as well as custom researcher-developed architectures.
+Here, we introduce ``FixedPointFinder``, an open-source Tensorflow toolbox for finding fixed points and linearized dynamics in arbitrary RNN architectures. ``FixedPointFinder`` leverages Tensorflow's efficient automatic differentiation back-end to perform the optimizations and Jacobian calculations required for fixed-point identification and characterization. ``FixedPointFinder`` works with any arbitrary RNN architecture that conforms to Tensorflow's ``RNNCell`` API, including popular gated architectures (e.g., long-short term memory (LSTM) or gated recurrent units (GRU)), traditional "vanilla" architectures, as well as custom researcher-developed architectures.
 
 # Example
 
 ``FixedPointFinder`` includes an end-to-end example that trains a Tensorflow RNN to solve a task and then identifies and visualizes the fixed points of the trained RNN. The task is the "flip-flop" task previously described in @SussilloBarak2013. Briefly, the task is to implement a 3-bit binary memory, in which each of 3 input channels delivers signed transient pulses (-1 or +1) to a corresponding bit of the memory, and an input pulse flips the state of that memory bit (also -1 or +1) whenever a pulse's sign is opposite of the current state of the bit. The example trains a 16-unit LSTM RNN to solve this task (Fig. 1). Once the RNN is trained, the example uses ``FixedPointFinder`` to identify and characterize the trained RNN's fixed points. Finally, the example produces a visualization of these results (Fig. 2). In addition to demonstrating a working use of ``FixedPointFinder``, this example provides a testbed for experimenting with different RNN architectures (e.g., numbers of recurrent units, LSTMs vs. GRUs vs. vanilla RNNs) and characterizing how these lower-level model design choices manifest in the higher-level dynamical implementation used to solve a task.
 
 ---
[…]
