### 2018-09-18 8c56753
Update paper.md
