### 2018-09-18 3d04fe9
Attempting to get figures to display in pdf
