### 2018-09-18 #962 @mattgolub
@arokem I'm having trouble getting both of the figures to display in the acticle proof pdf. Are you able to help with this? Both figures appear on the Github preview of the paper.md file.
