### 2018-09-18 c3ea9a3
Updated image links
