### 2018-09-19 1f873fd
Fixing paper compilation
