### 2018-09-19 #962 @arfon
@mattgolub - I've fixed the paper in https://github.com/mattgolub/fixed-point-finder/pull/1 . The `---` lines seemed to be suppressing the figure captions.
