### 2018-09-23 #962 @arokem
@mattgolub : happy to edit this submission. Looks interesting!

While I look around the submissions, do you have any suggestions for reviewers?
