### 2018-09-23 #962 @arokem
@mattgolub: Could you please clarify the relationship between this software and "recurrent whisperer"? Is it really an independent software library?
