### 2018-09-25 #962 @mattgolub
@arokem RecurrentWhisperer is another repository with code we have developed for training recurrent neural networks (RNNs) using Tensorflow. It is used to train the RNN used in the example that we provide, but it is not required (in general) to use FixedPointFinder.

FixedPointFinder is a tool meant for studying an RNN. The RNN to be studied will be different for each user--some may study RNNs they have trained, and others may study untrained (randomized) networks. FixedPointFinder is appropriate for all of these cases. In the example, we study of a network that we train using RecurrentWhisperer.

From a software engineering, it makes sense to keep RecurrentWhisperer as a separate repository, as it is used for other projects. However, if necessary for this submission, we can consider adding a copy of it into FixedPointFinder.
