### 2018-09-25 #962 @mattgolub
@arokem Are you still in need of reviewer suggestions? What are the criteria of a good reviewer candidate? Would it be the same as for a reviewer in our field-specific journals (computational neuroscience)?
