### 2018-09-30 #962 @arokem
Hi @mattgolub : thanks for the clarification. That makes sense.

As for reviewer: yes, if you have any suggestions for reviewers, that would be helpful. Maybe the author of similar software that you use yourself? It has to be someone who understands the development process and is able to weigh in on the organization of the software for use by others. Domain expertise is also good, but I would say that it's secondary. I'll look for someone as well.
