### 2018-09-30 #962 @arokem
👋 Hey @hiendn! Would you be willing to review this article for JOSS?
