### 2018-10-02 #962 @hiendn
> 👋 Hey @hiendn! Would you be willing to review this article for JOSS?

Hi @arokem. Unfortunately I am a bit overloaded at the moment. I will have to pass on this one :(
