### 2018-10-03 #962 @mattgolub
Hi @arokem. We can suggest Eric Jang (ejang@google.com) as a reviewer. He is an engineer at Google Brain with neuroscience interests.

Is one reviewer sufficient here? We can continue racking our brains for others if necessary.
