### 2018-10-03 1846439
Fixed bug in sequential optimizer
