### 2018-10-05 #962 @arokem
Thanks for the suggestion. I would rather try to find someone who doesn't share an affiliation with one of the authors, if possible.

@malmaud: would you be willing to review this article for JOSS?
