### 2018-10-05 b699a9b
addressed problem using pfor.batch_jacobian
