### 2018-10-07 #1003 @whedon
[REVIEW]: FixedPointFinder: A Tensorflow toolbox for identifying and characterizing fixed points in recurrent neural networks
**Submitting author:** @mattgolub (<a href="http://orcid.org/0000-0003-4508-0537">Matthew Golub</a>)
**Repository:** <a href="https://github.com/mattgolub/fixed-point-finder" target ="_blank">https://github.com/mattgolub/fixed-point-finder</a>
**Version:** v1.0
