### 2018-10-08 #1003 @malmaud
Hi @mattgolub, looking forward to reviewing this. Before digging deeper, you seem to be missing these items from the JOSS requirements:

* *Version: Does the release version given match the GitHub release (v1.0)?* - There is no tagged release on the project's Github page (https://github.com/mattgolub/fixed-point-finder/releases)
* *Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?* - I don't see any tests.
* *Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support* - I don't see any mention of guidelines in either the Readme or as a separate file.
