### 2018-10-08 7103590
Updated usage of pfor.batch_jacobian
