### 2018-10-08 fe0e461
Changed default setting of do_rerun_outliers to False
