### 2018-10-09 f0275e6
Removed superfluous fp_dict
