### 2018-10-09 1ab92cb
fixed import bug--previously relied on an import in the example
