### 2018-10-10 2765475
moved LSTMStateTuple utilities from FixedPointFinder.py to new tf_utils.py
