### 2018-10-10 bb6738e
removed LSTMStateTuple utilities from FixedPointFinder.py
