### 2018-10-10 b44d6a6
Removed obsolete gradient-based Jacobian functions
