+++ FixedPointFinder.py
@@ -1054,150 +1054,3 @@ class FixedPointFinder(object):
             ax.plot(z[:, 0], z[:, 1], **kwargs)
         elif n_dims == 1:
             ax.plot(z, **kwargs)
-
-    # ******************************************************************
-    # Previous Jacobian implementations, now obsolete w/ TF 1.10 *******
-    # ***** Can be re-integrated by users who require using earlier ****
-    # ***** versions of TensorFlow *************************************
-    # ******************************************************************
-
-    def _compute_single_jacobian_np_old(self, state_np):
-        '''Computes the Jacobian of the RNN state transition function at a
-        single point.
-
-        Uses 'brute force' assembling of tf.gradient calls. Very slow.
-
-        Args:
-            state_np: A [1 x n_dims] numpy array representing the point at
-            which to compute the Jacobian.
-
-        Returns:
-            An [n_dims x n_dims] numpy array containing the Jacobian of the
-            RNN state transition function at state_np.
-        '''
-
-        x, F, state, new_state = self._grab_RNN(state_np, self.inputs)
-        J = self._compute_single_jacobian_tf(F, x)
-
-        return J
-
-    def _compute_single_jacobian_np_new(self, state_np):
-        '''Computes the Jacobian of the RNN state transition function at a
-        single point.
-
-        Uses TF implementation of Jacobian computation, and is somewhat faster
-        than _compute_single_jacobian_np_old (about 50% faster on small toy
-        problems).
-
-        Args:
-            state_np: A [1 x n_dims] numpy array representing the point at
-            which to compute the Jacobian.
-
-        Returns:
-            An [n_dims x n_dims] numpy array containing the Jacobian of the
-            RNN state transition function at state_np.
-        '''
-
-        x, F, state, new_state = self._grab_RNN(state_np, self.inputs)
-        J_tf = tf.reshape(pfor.jacobian(F, x), [state_np.size, state_np.size])
-
-        return self.session.run(J_tf)
-
-    def _compute_single_jacobian_tf(self, F, x):
-        '''Computes the Jacobian of F at a single point, x.
-
-        Args:
-            F: A [1 x n_dims] tf.Tensor
-            x: A [1 x n_dims] tf.Variable.
-
-        Returns:
-            An [n_dims x n_dims] numpy array containing the Jacobian of F at x.
-
-        Raises:
-            ValueError: You should be calling _compute_multiple_jacobians_tf.
-        '''
-
-        n_inits, n_dims = x.shape
-
-        if n_inits > 1:
-            raise ValueError('You should be calling \
-                             _compute_multiple_jacobians_tf.')
-
-        print('\tComputing Jacobian at fixed point.')
-
-        # Assemble list of TF ops that will each compute the gradient
-        # of one dimension of the RNN function w.r.t. each dimension
-        # of the state
-        J_tf = []
-        for i in range(n_dims):
-            J_tf.append(tf.gradients(F[0, i], x)[0])
-
-        # Run that list of TF ops
-        J_tf_eval = self.session.run(J_tf)
-
-        # Assemble the evaluated TF ops into a numpy array
-        J_np = np.zeros([n_dims, n_dims])
-        for dim_idx in range(n_dims):
-            J_np[dim_idx, :] = J_tf_eval[dim_idx]
-
-        return J_np
-
-    def _compute_multiple_jacobians_tf(self, F, x):
-        '''Computes the Jacobian of F at multiple points, x.
-
-        Args:
-            F: An [n_inits x n_dims] tf.Tensor
-            x: An [n_inits x n_dims] tf.Variable
-
-        Returns:
-            An [n_inits x n_dims x n_dims] numpy array containing the Jacobian
-            of F at the points in x.
[…]
