### 2018-10-10 e035ffc
Bug fix: n_iters was overwritten with what is not unique_n_iters
