### 2018-10-10 820db71
added is_lstm()
