### 2018-10-10 142ee53
Added contribution guidelines
