+++ CONTRIBUTING.md
@@ -1,0 +1,6 @@
+## Contribution Guidelines
+
+Constructive contributions are welcome.
+- Questions, feedback, bug reports, and proposed features should be submitted as a GitHub [issue](https://github.com/mattgolub/fixed-point-finder/issues).
+- Alternatively, contact the repository owner, Matt, via email (mgolub@stanford.edu).
+- For development contributions, please first open an [issue](https://github.com/mattgolub/fixed-point-finder/issues) describing the proposed development. The resulting discussion may help prevent duplication of efforts. If moving forward with the development, open a pull request with the updated code or new features. Please reference the corresponding issue in the pull request.
