### 2018-10-10 7564bc7
Removed version info from header. GitHub will manage versioning henceforth.
