+++ example/FlipFlop.py
@@ -1,6 +1,5 @@
 '''
 flipflop.py
-Version 1.0
 Written using Python 2.7.12
 @ Matt Golub, August 2018.
 Please direct correspondence to mgolub@stanford.edu.
@@ -515,4 +514,3 @@ class FlipFlop(RecurrentWhisperer):
                 linestyle='--')
 
         plt.xlim(-1, n_time)
-
