### 2018-10-10 #1003 @mattgolub
Hi @malmaud, thank you for taking up this review and informing us of these missing items. Here is how we have worked to address these issues:

- *Version*: I've released v1.0. You can find the release [here](https://github.com/mattgolub/fixed-point-finder/releases/tag/v1.0).
- *Automated tests*: We had thought of the end-to-end example as a test and living documentation of the toolbox. Because you have brought this to our attention, we are also working to include a more formal testing suite that will run the FixedPointFinder and compare its results to ground truth results. This will be released by the end of the week, along with a corresponding update to the README.
- *Community guidelines*: I've added a separate file describing [contribution guidelines](https://github.com/mattgolub/fixed-point-finder/blob/master/CONTRIBUTING.md). We have added a reference to these guidelines at the end of the [README](https://github.com/mattgolub/fixed-point-finder/blob/master/README.md).
