### 2018-10-11 02d9986
added safe_shape and safe_index
