### 2018-10-11 40880eb
Major update: integration with new FixedPoints class and substantial improvements to the inner plumbing.
