+++ FixedPoints.py
@@ -1,0 +1,607 @@
+'''
+FixedPoints Class
+Supports FixedPointFinder
+Written using Python 2.7.12 and TensorFlow 1.10.
+@ Matt Golub, October 2018.
+Please direct correspondence to mgolub@stanford.edu.
+'''
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from sklearn.decomposition import PCA
+import matplotlib.pyplot as plt
+from mpl_toolkits.mplot3d import Axes3D
+
+import pdb
+import numpy as np
+import tf_utils
+
+class FixedPoints(object):
+    '''
+    A class for storing fixed points and associated data.
+    '''
+
+    def __init__(self,
+                 xstar=None,
+                 x_init=None,
+                 inputs=None,
+                 F_xstar=None,
+                 qstar=None,
+                 dq=None,
+                 n_iters=None,
+                 J_xstar=None,
+                 do_alloc_nan=False,
+                 n=None,
+                 n_states=None,
+                 n_inputs=None,
+                 tol_unique=1e-3,
+                 dtype=np.float32):
+        '''
+        Initializes a FixedPoints object with all input arguments as class
+        properties.
+
+        Optional args:
+
+            xstar: [n x n_states] numpy array with row xstar[i, :]
+            specifying an the fixed point identified from x_init[i, :].
+            Default: None.
+
+            x_init: [n x n_states] numpy array with row x_init[i, :]
+            specifying the initial state from which xstar[i, :] was optimized.
+            Default: None.
+
+            inputs: [n x n_inputs] numpy array with row inputs[i, :]
+            specifying the input to the RNN during the optimization of
+            xstar[i, :]. Default: None.
+
+            F_xstar: [n x n_states] numpy array with F_xstar[i, :]
+            specifying RNN state after transitioning from the fixed point in
+            xstar[i, :]. If the optimization succeeded (e.g., to 'tol') and
+            identified a stable fixed point, the state should not move
+            substantially from the fixed point (i.e., xstar[i, :] should be
+            very close to F_xstar[i, :]). Default: None.
+
+            qstar: [n,] numpy array with qstar[i] containing the
+            optimized objective (1/2)(x-F(x))^T(x-F(x)), where
+            x = xstar[i, :]^T and F is the RNN transition function (with the
+            specified constant inputs). Default: None.
+
+            dq: [n,] numpy array with dq[i] containing the absolute
+            difference in the objective function after (i.e., qstar[i]) vs
+            before the final gradient descent step of the optimization of
+            xstar[i, :]. Default: None.
+
+            n_iters: [n,] numpy array with n_iters[i] as the number of
+            gradient descent iterations completed to yield xstar[i, :].
+            Default: None.
+
+            J_xstar: [n x n_states x n_states] numpy array with
+            J_xstar[i, :, :] containing the Jacobian of the RNN state
+            transition function at fixed point xstar[i, :]. Default: None,
+            which results in an appropriately sized numpy array of NaNs.
+            Default: None.
+
+            do_alloc_nan: Bool indicating whether to initialize all data attributes (all optional args above) as NaN-filled numpy arrays.
+            If True, n, n_states and n_inputs must be provided. These values
+            are otherwise ignored:
+
+                n: Positive int specifying the number of fixed points to
+                allocate space for.
+
+                n_states: Positive int specifying the dimensionality of the
+                network state (a.k.a. the number of hidden units).
+
+                n_inputs: Positive int specifying the dimensionality of the
+                network inputs.
+
+            tol_unique: Positive scalar specifying the numerical
[…]
