### 2018-10-11 f960edb
Major update: integration with new FixedPoints class and substantial improvements to the inner plumbing.
