+++ FixedPointFinder.py
@@ -1,7 +1,7 @@
 '''
 TensorFlow FixedPointFinder
 Written using Python 2.7.12 and TensorFlow 1.10.
-@ Matt Golub, September 2018.
+@ Matt Golub, October 2018.
 Please direct correspondence to mgolub@stanford.edu.
 '''
 
@@ -17,12 +17,9 @@ import tensorflow as tf
 from tensorflow.python.ops import parallel_for as pfor
 import absl
 
-from sklearn.decomposition import PCA
-import matplotlib.pyplot as plt
-from mpl_toolkits.mplot3d import Axes3D
-
 PATH_TO_RECURRENT_WHISPERER = '../../recurrent-whisperer/'
 sys.path.insert(0, PATH_TO_RECURRENT_WHISPERER)
+from FixedPoints import FixedPoints
 from AdaptiveLearningRate import AdaptiveLearningRate
 from AdaptiveGradNormClip import AdaptiveGradNormClip
 import tf_utils
@@ -34,9 +31,10 @@ class FixedPointFinder(object):
         max_iters=5000,
         method='joint',
         do_rerun_outliers=False,
-        outlier_q_scale=1000,
+        outlier_q_scale=10,
         tol_unique=1e-3,
         do_compute_jacobians=True,
+        tf_dtype=tf.float32,
         verbose=False,
         alr_hps=dict(),
         agnc_hps=dict(),
@@ -84,6 +82,10 @@ class FixedPointFinder(object):
             do_compute_jacobians (optional): A bool specifying whether or not
             to compute the Jacobian at each fixed point. Default: True.
 
+            tf_dtype: Data type to use for all TensorFlow ops. The
+            corresponding numpy data type is used for numpy objects and
+            operations.
+
             verbose (optional): A bool specifying whether or not to print
             per-iteration updates during each optimization. Default: False.
 
@@ -105,15 +107,11 @@ class FixedPointFinder(object):
 
         self.rnn_cell = rnn_cell
         self.session = sess
+        self.tf_dtype = tf_dtype
+        self.np_dtype = tf_dtype.as_numpy_dtype
 
         self.is_lstm = isinstance(
             rnn_cell.state_size, tf.nn.rnn_cell.LSTMStateTuple)
-
-        # These class variables are set by find_fixed_points
-        self.initial_states = None
-        self.inputs = None
-        self.n_inits = None
-        self.n_dims = None
 
         # *********************************************************************
         # Optimization hyperparameters ****************************************
@@ -131,21 +129,6 @@ class FixedPointFinder(object):
         self.grad_norm_clip_hps = agnc_hps
         self.adam_optimizer_hps = adam_hps
 
-        ''' These class variables are set either by _run_joint_optimization or
-        _run_sequential_optimizations'''
-        self.xstar = None
-        self.F_xstar = None
-        self.qstar = None
-        self.dq = None
-
-        ''' These class variables are set by _identify_unique_fixed_points and
-        _compute_jacobians_at_unique_fixed_points'''
-        self.unique_xstar = None
-        self.unique_Fxstar = None
-        self.unique_J_xstar = None
-        self.unique_q = None
-        self.unique_dq = None
-
     def sample_states(self, state_traj, n_inits,
                       noise_scale=0.0, rng=npr.RandomState(0)):
         '''Draws random samples from trajectories of the RNN state. Samples
@@ -154,9 +137,9 @@ class FixedPointFinder(object):
         states for fixed point optimizations.
 
         Args:
-            state_traj: [n_batch x n_time x n_dims] numpy array or
-            LSTMStateTuple with .c and .h as [n_batch x n_time x n_dims] numpy
-            arrays. Contains example trajectories of the RNN state.
[…]
