### 2018-10-11 8dd55f3
Uniqueness check now includes inputs
