### 2018-10-11 a8fee6d
Uniqueness check now includes inputs
