### 2018-10-11 9b80517
Updated printing statements
