### 2018-10-12 e9c440b
Path updates now done via bash (export) rather than in .py files
