+++ FixedPointFinder.py
@@ -10,7 +10,6 @@ from __future__ import division
 from __future__ import print_function
 
 import pdb
-import sys
 import numpy as np
 import numpy.random as npr
 import tensorflow as tf
