### 2018-10-12 mattgolub/fixed-point-finder#3 @malmaud
Consider requirements.txt file
Since all the dependencies seem to be installable via `pip`, you can create a [requirements.txt](https://pip.pypa.io/en/stable/user_guide/#requirements-files) file that lists them to make it easy for others to install the dependencies.
