### 2018-10-12 #1003 @malmaud
Everything seems good - the package installed and the included example runs successfully on my machine. It's clearly solving an important task relevant to many in the ML community. The article is clear and well-written. The codebase is small but non-trivial. The public API methods have good docstrings attached.

I think just a few things and it's ready for acceptance:

* As mentioned earlier, some kind of tests are needed. Doesn't have to be anything too fancy.
* The installation is a little complex. You have to clone this package, clone a package it depends on, manually add both to your Python path, manually look through the README to see the other Python packages it depends on, and then manually install those with pip.

I'd like to see if that can be simplified a bit - either this package can have a `setup.py` so it itself is installable with pip, or the packages it depends on could be listed in a requirements.txt file, or a Dockerfile could be included that bundles everything up, etc.

* It's Python 2 only at the moment. That's not a deal breaker necessarily, although it's unfortunate given the increasing use of Python 3 and I think it would take just a small amount amount of work to make it Python 3 compatible given the small amount of code involved (the package is on the order of 2000 lines of Python).

cf
* https://github.com/mattgolub/fixed-point-finder/issues/3
* https://github.com/mattgolub/fixed-point-finder/issues/2
