### 2018-10-13 5b43e3c
Major efficiency improvements; Added save/load functionality to FixedPoints class
