+++ FixedPointFinder.py
@@ -215,9 +215,10 @@ class FixedPointFinder(object):
 
         if inputs.shape[0] == 1:
             inputs_nxd = np.tile(inputs, [n, 1]) # safe, even if n == 1.
-        else:
-            # assumes inputs.shape[0] == n. should check.
+        elif inputs.shape[0] == n:
             inputs_nxd = inputs
+        else:
+            raise ValueError('Incompatible inputs shape: %s.' % inputs.shape)
 
         if self.method == 'sequential':
             all_fps = self._run_sequential_optimizations(
@@ -229,10 +230,17 @@ class FixedPointFinder(object):
             raise ValueError('Unsupported optimization method. Must be either \
                 \'joint\' or \'sequential\', but was  \'%s\'' % self.method)
 
-        if self.do_rerun_outliers:
-            all_fps = self._run_additional_iterations_on_outliers(all_fps)
-
+        # Filter out duplicates after from the first optimization round
         unique_fps = all_fps.get_unique()
+
+        # Optionally run additional optimization iterations on identified
+        # fixed points with q values on the large side of the q-distribution.
+        if self.do_rerun_outliers:
+            unique_fps = self._run_additional_iterations_on_outliers(
+                unique_fps)
+
+            # Filter out duplicates after from the second optimization round
+            unique_fps = unique_fps.get_unique()
 
         if self.do_compute_jacobians:
             print('Computing Jacobian at %d '
@@ -263,47 +271,18 @@ class FixedPointFinder(object):
             fps: A FixedPoints object containing the optimized fixed points
             and associated metadata.
         '''
+        print('\nFinding fixed points via joint optimization...')
+
         n, _ = tf_utils.safe_shape(initial_states)
 
         x, F = self._grab_RNN(initial_states, inputs)
 
-        # an array of objectives (one per initial state) to be combined below
-        q_1xn = 0.5 * tf.reduce_sum(tf.square(F - x), axis=1)
+        # A shape [n,] TF Tensor of objectives (one per initial state) to be
+        # combined in _run_optimization_loop.
+        q = 0.5 * tf.reduce_sum(tf.square(F - x), axis=1)
 
-        '''There are two obvious choices of how to combine multiple objectives
-        here: minimizing the maximum value; or minimizing the mean value.
-        While the former allows for direct checks for convergence for each
-        fixed point, the latter is empirically much more efficient (more
-        progress made in fewer gradient steps).
-
-
-        max: This should have nonzero gradients only for the state with the
-        largest q. If so, in effect, this will wind up doing a sequential
-        optimization.
-
-        mean: This should make progress on many of the states at each step,
-        which likely speeds things up. However, one could imagine pathological
-        situations arising where the objective continues to improve due to
-        improvements in some fixed points but not others.'''
-
-        q = tf.reduce_mean(q_1xn)
-
-        print('\nFinding fixed points via joint optimization...')
-        xstar, F_xstar, _, dq, n_iters = self._run_optimization_loop(q, x, F)
-
-        '''Replace mean qstar (the scalar loss function) with the individual q
-        values at each fixed point'''
-        qstar = self.session.run(q_1xn)
-
-        '''Note, there is no meaningful way to get individual dq values
-        without following up with at least one gradient step from the
-        sequential optimizer for each fixed point. This is likely a good idea
-        to do anyways, since it would allow fine tuning at slow points that
-        are not local minima, but do not prevent the joint optimization from
-        reaching termination criteria.'''
-
-        dq = dq * np.ones([n])
[…]
