### 2018-10-13 055b10b
Added methods: find, __contains__
