### 2018-10-13 5fab5be
Updated printing
