### 2018-10-14 fa17b43
Update setup.py
