### 2018-10-14 #1003 @arokem
I would strongly urge an update to Python 3. If you want others to use the software into the future, and combine it with other packages in the scientific Python eco-system that would be a good idea. Many major packages are going to stop supporting Python 2 as soon as just over a year from now (https://python3statement.org/). The publication of this paper seems like a good time to undertake this update.
