### 2018-10-15 7dda434
Added fixed-point-finder/ to python path
