### 2018-10-15 dcfdec1
Added requirements.txt files
