+++ requirements-gpu.txt
@@ -1,0 +1,7 @@
+numpy==1.14.5
+scipy==1.1.0
+scikit-learn==0.20.0
+matplotlib==2.2.3
+PyYAML==3.13
+tensorflow-gpu==1.10.0
+-e git://github.com/mattgolub/recurrent-whisperer.git@v1.0.0#egg=recurrent-whisperer
