### 2018-10-15 ff80741
Update install instructions; Update usage to reflect FixedPoints redesign.
