+++ README.md
@@ -2,48 +2,72 @@
 
 This code finds and analyzes the fixed points of recurrent neural networks that have been built using Tensorflow. The approach follows that outlined in Sussillo and Barak, "Opening the Black Box: Low-Dimensional Dynamics in High-Dimensional Recurrent Neural Networks", Neural Computation 2013 ([link](https://doi.org/10.1162/NECO_a_00409)).
 
-## Prerequisites
+Written using Python 2.7.12.
 
-The code is written in Python 2.7.6. You will also need:
+## Recommended Installation
 
-* **TensorFlow** version 1.10 ([install](https://www.tensorflow.org/install/)).
+1. [Clone or download](https://help.github.com/articles/cloning-a-repository/) this repository.
+2. Create a virtual environment for the required dependencies:
+    To create a new virtual environment specific to Python 2.7, enter at the command line:
+    ```bash
+    $ virtualenv --system-site-packages -p python2.7 your-virtual-env-name
+    ```
+    where `your-virtual-env-name` is a path to the the virtual environment you would like to create (e.g.: `/home/fpf`). Then   activate your new virtual environment:
+    ```bash
+    $ source your-virtual-env-name/bin/activate
+    ```
+    When you are finished working in your virtual environment (not now), enter:
+    ```bash
+    $ deactivate
+    ```
+3. Automatically assemble all dependencies using `pip` and the `requirements*.txt` files. This step will depend on whether you require Tensorflow with GPU support.
+
+    For GPU-enabled TensorFlow, use:
+
+    ```bash
+    $ pip install -r requirements-gpu.txt
+    ```
+
+    For CPU-only TensorFlow, use:
+
+    ```bash
+    $ pip install -r requirements-cpu.txt
+    ```
+
+## Advanced Installation
+
+Advanced Python users may skip the Recommended Installation, opting to instead clone this repository and ensure that compatible versions of the following prerequisites are available:
+
+* **TensorFlow** (requires at least version 1.10) ([install](https://www.tensorflow.org/install/)).
 * **NumPy, SciPy, Matplotlib** ([install SciPy stack](https://www.scipy.org/install.html), contains all of them).
 * **Scikit-learn** ([install](http://scikit-learn.org/)).
 * **PyYaml** ([install](https://pyyaml.org/)).
-* **RecurrentWhisperer** ([clone or download](https://github.com/mattgolub/recurrent-whisperer/)); see "Installation" below.
-
-
-## Installation
-
-1. Install the prerequisite packages listed above.
-2. Clone or download the [FixedPointFinder repository](https://help.github.com/articles/cloning-a-repository/).
-4. Add the directories for ```FixedPointFinder``` and ```RecurrentWhisperer```  to your Python path:
-```bash
-$ export PYTHONPATH=$PYTHONPATH:/path/to/your/directory/fixed-point-finder/
-$ export PYTHONPATH=$PYTHONPATH:/path/to/your/directory/recurrent-whisperer/
-```
-where "/path/to/your/directory" is replaced with the path to the corresponding repository. This step must be performed each time you launch a new terminal to work with ```FixedPointFinder```, and thus you may want to add the lines above to a startup script (e.g., the .bashrc / .bashprofile script in your home folder or an activate script in your virtual environment).
+* **RecurrentWhisperer** ([install](https://github.com/mattgolub/recurrent-whisperer/)).
 
 ## General Usage
 
 1. Start by building, and if desired, training an RNN. ```FixedPointFinder``` works with any arbitrary RNN that conforms to Tensorflow's `RNNCell` API.
 2. Build a ```FixedPointFinder``` object:
-  ```python
-  >>> fpf = FixedPointFinder(your_rnn_cell, tf_session, **hyperparams)
-  ```
[…]
