### 2018-10-15 mattgolub/fixed-point-finder#3 @mattgolub
@malmaud  Thanks for the great suggestion. I've created requirements.txt files and updated the installation instructions in the README.
