### 2018-10-15 df164de
Updated verbosity choices
