+++ FixedPointFinder.py
@@ -32,7 +32,8 @@ class FixedPointFinder(object):
         tol_unique=1e-3,
         do_compute_jacobians=True,
         tf_dtype=tf.float32,
-        verbose=False,
+        verbose=True,
+        super_verbose=False,
         alr_hps=dict(),
         agnc_hps=dict(),
         adam_hps={'epsilon': 0.01}):
@@ -83,7 +84,10 @@ class FixedPointFinder(object):
             corresponding numpy data type is used for numpy objects and
             operations.
 
-            verbose (optional): A bool specifying whether or not to print
+            verbose (optional): A bool indicating whether to print high-level
+            status updates. Default: True.
+
+            super_verbose (optional): A bool indicating whether or not to print
             per-iteration updates during each optimization. Default: False.
 
             alr_hps (optional): A dict containing hyperparameters governing an
@@ -121,6 +125,7 @@ class FixedPointFinder(object):
         self.tol_unique = tol_unique
         self.do_compute_jacobians = do_compute_jacobians
         self.verbose = verbose
+        self.super_verbose = super_verbose
 
         self.adaptive_learning_rate_hps = alr_hps
         self.grad_norm_clip_hps = agnc_hps
@@ -213,7 +218,8 @@ class FixedPointFinder(object):
         '''
         n = tf_utils.safe_shape(initial_states)[0]
 
-        print('\nSearching for fixed points from %d initial states.\n' % n)
+        self._print_if_verbose('\nSearching for fixed points '
+                               'from %d initial states.\n' % n)
 
         if inputs.shape[0] == 1:
             inputs_nxd = np.tile(inputs, [n, 1]) # safe, even if n == 1.
@@ -235,7 +241,8 @@ class FixedPointFinder(object):
         # Filter out duplicates after from the first optimization round
         unique_fps = all_fps.get_unique()
 
-        print('\tIdentified %d unique fixed points.' % unique_fps.n)
+        self._print_if_verbose('\tIdentified %d unique fixed points.' %
+                               unique_fps.n)
 
         # Optionally run additional optimization iterations on identified
         # fixed points with q values on the large side of the q-distribution.
@@ -247,12 +254,12 @@ class FixedPointFinder(object):
             unique_fps = unique_fps.get_unique()
 
         if self.do_compute_jacobians:
-            print('\tComputing Jacobian at %d '
-                  'unique fixed points.' % unique_fps.n)
+            self._print_if_verbose('\tComputing Jacobian at %d '
+                                   'unique fixed points.' % unique_fps.n)
             J_xstar = self._compute_multiple_jacobians_np(unique_fps)
             unique_fps.J_xstar = J_xstar
 
-        print('\n\tFixed point finding complete.\n')
+        self._print_if_verbose('\n\tFixed point finding complete.\n')
 
         return unique_fps, all_fps
 
@@ -275,7 +282,8 @@ class FixedPointFinder(object):
             fps: A FixedPoints object containing the optimized fixed points
             and associated metadata.
         '''
-        print('\tFinding fixed points via joint optimization.')
+        self._print_if_verbose('\tFinding fixed points '
+                               'via joint optimization.')
 
         n, _ = tf_utils.safe_shape(initial_states)
 
@@ -331,7 +339,8 @@ class FixedPointFinder(object):
         is_fresh_start = q_prior is None
 
         if is_fresh_start:
-            print('\tFinding fixed points via sequential optimizations...')
+            self._print_if_verbose('\tFinding fixed points via '
+                                   'sequential optimizations...')
 
         n_inits, n_states = tf_utils.safe_shape(initial_states)
         n_inputs = inputs.shape[1]
@@ -351,10 +360,10 @@ class FixedPointFinder(object):
             inputs_i = inputs[index, :]
 
             if is_fresh_start:
[…]
