### 2018-10-15 887bdb7
Added tests
