+++ test/test_utils.py
@@ -1,0 +1,152 @@
+'''
+test_utils.py
+Written using Python 2.7.12
+@ Matt Golub, October 2018.
+Please direct correspondence to mgolub@stanford.edu.
+'''
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+import numpy as np
+import tensorflow as tf
+import matplotlib.pyplot as plt
+
+def build_test_rnn(n_hidden, n_inputs, session):
+    '''Build a "vanilla" RNNCell and deterministically set its weights. The
+    RNN's fixed points are thus also deterministic, allowing ground truth
+    values to be compared against those found by FixedPointFinder.
+
+    Args:
+        n_hidden:
+            Non-negative int specifying the number of hidden units in the RNN.
+
+        n_inputs:
+            Non-negative int specifying the number of inputs to the RNN.
+            Inputs are ignored by these test RNN's, so this value can be set
+            arbitraily without affecting results. This is included only to
+            ensure input dimensionality is consistent across the various objects used in a test.
+
+        session:
+            A Tensorflow session within which to initialize the RNNCell.
+    '''
+    rnn_cell = tf.nn.rnn_cell.BasicRNNCell(n_hidden)
+    # rnn_cell = tf.nn.rnn_cell.GRUCell(n_hidden)
+    # rnn_cell = tf.nn.rnn_cell.LSTMCell(n_hidden)
+
+    # These will stay as 0's
+    W_np_input = np.zeros([n_inputs, n_hidden], dtype=np.float32)
+    b_np = np.zeros([n_hidden], dtype=np.float32)
+
+    # This will be set deterministically
+    W_np_state = np.zeros([n_hidden, n_hidden], dtype=np.float32)
+
+    '''The following was chosen to such that resulting networks have some non-
+    origin fixed points.'''
+    row_vals = np.linspace(-1, 1, n_hidden)
+    for in_idx in range(n_hidden):
+        W_np_state[in_idx, :] = -np.roll(row_vals, -in_idx)
+    # print('\nW=' + str(W_np_state))
+
+    W_np = np.concatenate((W_np_input, W_np_state), axis=0)
+
+    input_data = tf.Variable(np.zeros([1, n_inputs]), dtype=tf.float32)
+    state = tf.Variable(np.zeros([1, n_hidden]), dtype=tf.float32)
+
+    output, final_state = rnn_cell(input_data, state)
+    W_tf, b_tf = rnn_cell.variables
+
+    assign_W = tf.assign(W_tf, W_np)
+    assign_b = tf.assign(b_tf, b_np)
+
+    session.run(tf.global_variables_initializer())
+    session.run([assign_W, assign_b])
+
+    return rnn_cell
+
+def generate_initial_states_and_inputs(n_hidden, n_inputs,
+                                       n_inits_per_state_dim=3,
+                                       min_val_per_state_dim=-1.0,
+                                       max_val_per_state_dim=1.0,
+                                       debug=False):
+    '''Generates a grid of initial states and zeroed-out inputs in a format
+    compatible with RNNCell.
+
+    Args:
+        n_hidden:
+            Non-negative int specifying the number of hidden units in the RNN.
+
+        n_inputs:
+            Non-negative int specifying the number of inputs to the RNN.
+            Inputs are ignored by these test RNN's, so this value can be set
+            arbitraily without affecting results. This is included only to
+            ensure input dimensionality is consistent across the various
+            objects used in a test.
+
+        n_inits_per_state_dim (optional):
+            Non-negative int specifying the grid resolution for each
+            dimension. The total number of grid states will be
+            pow(n_inits_per_state_dim, n_hidden). Default: 3.
+
+        min_val_per_state_dim (optional):
[…]
