### 2018-10-15 #1003 @mattgolub
@malmaud Thank you again for your time and effort in this review. I believe I have addressed the needs that you laid out for publication with JOSS:

-*Installation*: I've added requirements*.txt files along with corresponding instructions for a more streamlined installation in the README. I've noted this in the GitHub issue on this topic, and I've closed the issue.

-*Automated Testing*: I've added a testing suite along with corresponding instructions in the README.

-*Python 3*: The rationale for developing using Python 2.7 has been that a large set of collaborators (for whom this tool is immediately intended; including many at Google Brain/AI) are still primarily developing using Python 2.7. I see the very likely future need for compatibility with Python 3. I appreciate that you noted that this is "not a deal breaker." If it's alright with you, I'll take care of this in future work (i.e., after the JOSS paper is out).
