### 2018-10-18 #1003 @malmaud
Thanks - all my objections have been addressed. The automated tests run to completion on my machine.

I'm personally OK with adding Python 3 support sometime later, but I think that will ultimately be up to the editor.

@arokem, I recommend accepting at this time, contingent on you being OK with the Python 3 issue.
