### 2018-10-22 c72008d
updated min_loss
