### 2018-10-24 adfed94
Updated example hps
