+++ example/run_FlipFlop.py
@@ -32,7 +32,7 @@ alr_hps = {'initial_rate': 0.1}
 hps = {
     'rnn_type': 'lstm',
     'n_hidden': 16,
-    'min_loss': 2e-4,
+    'min_loss': 1e-4,
     'min_learning_rate': 1e-5,
     'log_dir': './logs/',
     'data_hps': {
