### 2018-10-24 4702179
Updated requirements*.txt to point to RecurrentWhisperer v1.0.2
