+++ requirements-gpu.txt
@@ -4,4 +4,4 @@ scikit-learn==0.20.0
 matplotlib==2.2.3
 PyYAML==3.13
 tensorflow-gpu==1.10.0
--e git://github.com/mattgolub/recurrent-whisperer.git@v1.0.1#egg=recurrent-whisperer
+-e git://github.com/mattgolub/recurrent-whisperer.git@v1.0.2#egg=recurrent-whisperer
