### 2018-10-26 bdfb7ce
Renamed merged_opt_summaries
