### 2018-10-29 #1003 @arokem
Sorry for the slowness here.

@malmaud: thank you for the review! 👏

The paper is ready to be accepted, but ust one more thing, @mattgolub : could you please add DOIs for all the references in your bibliography? Once that's done, could you please create an archive for the software as it is after making all these changes (e.g., by tagging a release and uploading to zenodo.org). Once we have that, we can go ahead with the acceptance and archiving of the paper.
