### 2018-10-31 d67e9f6
Added DOIs
