+++ paper/paper.bib
@@ -6,6 +6,7 @@
   number={7474},
   pages={78},
   year={2013},
+  doi = {10.1038/nature12742},
   publisher={Nature Publishing Group}
 }
 
@@ -17,6 +18,7 @@
   number={5},
   pages={1005--1019},
   year={2018},
+  doi = {10.1016/j.neuron.2018.05.020},
   publisher={Elsevier}
 }
 
@@ -28,6 +30,7 @@
   number={3},
   pages={626--649},
   year={2013},
+  doi = {10.1162/NECO_a_00409},
   publisher={MIT Press}
 }
 
@@ -35,9 +38,12 @@
   title={Inferring single-trial neural population dynamics using sequential auto-encoders},
   author={Pandarinath, Chethan and O'Shea, Daniel J and Collins, Jasmine and Jozefowicz, Rafal and Stavisky, Sergey D and Kao, Jonathan C and Trautmann, Eric M and Kaufman, Matthew T and Ryu, Stephen I and Hochberg, Leigh R and Henderson, Jaimie M. and Shenoy, Krishna V. and Abbott, L.F. and Sussillo, D},
   journal={Nature Methods},
+  volume={15},
+  number={10},
+  pages={805--815},
   year={2018},
+  DOI={10.1038/s41592-018-0109-9},
   publisher={Nature Publishing Group},
-  DOI={10.1038/s41592-018-0109-9}
 }
 
 @article{Katz2018,
@@ -48,5 +54,6 @@
   number={8},
   pages={3636--3646},
   year={2018},
+  doi = {10.1109/TNNLS.2017.2733544},
   publisher={IEEE}
 }
