### 2018-10-31 5e9e46d
Prep for JOSS publication
