+++ paper/paper.md
@@ -32,7 +32,7 @@ Recurrent neural networks (RNNs) are powerful function approximators that can be
 
 Perhaps the most critical step in this reverse-engineering is identifying the fixed points of the RNN's internal dynamics. These fixed points, along with local linearizations of the RNN dynamics about those fixed points, provide a mechanistic description of how the network implements a computation. Identifying the fixed points of an RNN requires optimization tools that have previously been tailored for specific, "vanilla" or Hopfield-like RNN architectures [@SussilloBarak2013, @Katz2018]. These approaches rely on hard-coded analytic derivatives (e.g., of the hyperbolic tangent function) and thus can be cumbersome for complicated models (e.g., gated architectures), for studies that involve multiple different models, or when frequent architectural changes are required during the development of a model. While numerical derivatives can partially alleviate these difficulties, the additional computational costs typically make them infeasible for anything but the smallest of models.
 
-Here, we introduce ``FixedPointFinder``, an open-source Tensorflow toolbox for finding fixed points and linearized dynamics in arbitrary RNN architectures. ``FixedPointFinder`` leverages Tensorflow's efficient automatic differentiation back-end to perform the optimizations and Jacobian calculations required for fixed-point identification and characterization. ``FixedPointFinder`` works with any arbitrary RNN architecture that conforms to Tensorflow's ``RNNCell`` API, including popular gated architectures (e.g., long-short term memory (LSTM) or gated recurrent units (GRU)), traditional "vanilla" architectures, as well as custom researcher-developed architectures.
+Here, we introduce ``FixedPointFinder``, an open-source Tensorflow toolbox for finding fixed points and linearized dynamics in arbitrary RNN architectures. ``FixedPointFinder`` leverages Tensorflow's efficient automatic differentiation back-end to perform the optimizations and Jacobian calculations required for fixed-point identification and characterization. ``FixedPointFinder`` works with any arbitrary RNN architecture that conforms to Tensorflow's ``RNNCell`` API, including popular gated architectures (e.g., long short-term memory (LSTM) or gated recurrent units (GRU)), traditional "vanilla" architectures, as well as custom researcher-developed architectures.
 
 # Example
 
