### 2018-10-31 #1003 @mattgolub
@arokem As requested, I've added DOIs to the bibliography and created a Zenodo release (10.5281/zenodo.1476209). Thanks again for all of your efforts here.

@malmaud Thanks again for the review.
